output: Extract candidate preview summaries
Candidate preview output renders candidate overviews and numbered previews
in one module. The renderer also derives target summaries and snippet
models before printing human and porcelain output.

That couples summary derivation to terminal rendering. The output layer has
no named boundary for the candidate overview model, so tests and future
renderers have to reach through the renderer for summary state.

This commit extracts candidate target summaries, snippet lines, ambiguity
block summaries, and common target detection into
output.candidate_preview_summary while leaving rendering in
output.candidate_preview.

Subsequent commits will retarget candidate preview tests and architecture
checks to the new summary boundary.
